requirements.txt
asgi-redis==1.4.0
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.1
Automat==0.6.0
backports.shutil-get-terminal-size==1.0.0
certifi==2017.4.17
channels==1.1.3
chardet==3.0.4
configparser==3.5.0
constantly==15.1.0
coverage==4.3.4
daphne==1.2.0
decorator==4.0.11
Django==1.11.2
django-admin-tools==0.8.1
django-admin-tools-stats==0.8.0
django-bower==5.2.0
django-cache-utils==0.7.2
django-grappelli==2.10.1
django-jsonfield==1.0.1
django-nose==1.4.4
django-nvd3==0.9.7
django-qsstats-magic==0.7.2
djangorestframework==3.6.2
enum34==1.1.6
flake8==3.3.0
funcsigs==1.0.2
futures==3.1.1
greenlet==0.4.12
httpie==0.9.9
hyperlink==17.1.1
idna==2.5
incremental==17.5.0
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
msgpack-python==0.4.8
neovim==0.1.13
nose==1.3.7
pathlib2==2.2.1
pbr==3.0.1
pexpect==4.2.1
pickleshare==0.7.4
pluggy==0.4.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
python-dateutil==2.6.0
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2017.2
redis==2.10.5
requests==2.13.0
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
tox==2.7.0
traitlets==4.3.2
trollius==2.1
Twisted==17.5.0
txaio==2.8.0
Unidecode==0.4.20
urllib3==1.21.1
uWSGI==2.0.15
virtualenv==15.1.0
wcwidth==0.1.7
zope.interface==4.4.1
